requirements.txt
opencv-python-headless==4.7.0.72
pillow==9.5.0
pytesseract==0.3.10
pdf2image==1.16.3
pandas==2.2.2
numpy==1.26.4
ocrmypdf==13.5.0